# perf(rust, python): bytes row format for streaming groupby/unique keys `>3.5x`