This looks very promising. The first naive solution which doesn't reuse buffers yet already shows to be 3.7 times faster on a db-benchmark query. 

This query groups by 2 integer keys. I expect the benchmark on strings to perform even better as we elide the string indirection cache misses.

## main
```
 Performance counter stats for './bin/0_18_1 /home/ritchie46/code/db-benchmark/data/G1_1e8_1e2_0_0.ipc 12 true':

         23.026,27 msec task-clock                       #    9,969 CPUs utilized             
             1.486      context-switches                 #   64,535 /sec                      
               108      cpu-migrations                   #    4,690 /sec                      
           867.597      page-faults                      #   37,679 K/sec                     
    81.954.648.691      cpu_core/cycles/                 #    3,559 G/sec                       (39,44%)
    68.795.644.229      cpu_atom/cycles/                 #    2,988 G/sec                       (64,04%)
   107.438.026.937      cpu_core/instructions/           #    4,666 G/sec                       (39,44%)
    91.380.257.702      cpu_atom/instructions/           #    3,969 G/sec                       (64,04%)
    14.752.392.664      cpu_core/branches/               #  640,676 M/sec                       (39,44%)
    12.459.702.825      cpu_atom/branches/               #  541,108 M/sec                       (64,04%)
       106.063.183      cpu_core/branch-misses/          #    4,606 M/sec                       (39,44%)
        93.445.771      cpu_atom/branch-misses/          #    4,058 M/sec                       (64,04%)
   491.727.892.146      cpu_core/slots/                  #   21,355 G/sec                       (39,44%)
   109.368.087.119      cpu_core/topdown-retiring/       #     22,3% Retiring                   (39,44%)
    44.100.356.779      cpu_core/topdown-bad-spec/       #      9,0% Bad Speculation            (39,44%)
    19.172.138.703      cpu_core/topdown-fe-bound/       #      3,9% Frontend Bound             (39,44%)
   318.249.101.705      cpu_core/topdown-be-bound/       #     64,8% Backend Bound              (39,44%)
     7.531.842.616      cpu_core/topdown-heavy-ops/      #      1,5% Heavy Operations          #     20,7% Light Operations           (39,44%)
    19.338.115.002      cpu_core/topdown-br-mispredict/  #      3,9% Branch Mispredict         #      5,0% Machine Clears             (39,44%)
     8.532.895.399      cpu_core/topdown-fetch-lat/      #      1,7% Fetch Latency             #      2,2% Fetch Bandwidth            (39,44%)
   305.363.100.619      cpu_core/topdown-mem-bound/      #     62,2% Memory Bound              #      2,6% Core Bound                 (39,44%)

       2,309736890 seconds time elapsed

      21,821341000 seconds user
       1,207631000 seconds sys
```

## PR

```
 Performance counter stats for './bin/dev /home/ritchie46/code/db-benchmark/data/G1_1e8_1e2_0_0.ipc 12 true':

          5.730,44 msec task-clock                       #    9,156 CPUs utilized             
               485      context-switches                 #   84,636 /sec                      
                40      cpu-migrations                   #    6,980 /sec                      
           660.508      page-faults                      #  115,263 K/sec                     
    22.011.372.603      cpu_core/cycles/                 #    3,841 G/sec                       (38,85%)
    15.849.037.670      cpu_atom/cycles/                 #    2,766 G/sec                       (67,75%)
    56.962.131.133      cpu_core/instructions/           #    9,940 G/sec                       (38,85%)
    21.983.432.578      cpu_atom/instructions/           #    3,836 G/sec                       (67,75%)
     8.505.348.900      cpu_core/branches/               #    1,484 G/sec                       (38,85%)
     3.288.282.193      cpu_atom/branches/               #  573,827 M/sec                       (67,75%)
        27.987.163      cpu_core/branch-misses/          #    4,884 M/sec                       (38,85%)
         6.626.732      cpu_atom/branch-misses/          #    1,156 M/sec                       (67,75%)
   132.068.235.621      cpu_core/slots/                  #   23,047 G/sec                       (38,85%)
    56.565.527.163      cpu_core/topdown-retiring/       #     43,0% Retiring                   (38,85%)
     9.488.750.852      cpu_core/topdown-bad-spec/       #      7,2% Bad Speculation            (38,85%)
     5.650.653.088      cpu_core/topdown-fe-bound/       #      4,3% Frontend Bound             (38,85%)
    59.785.548.749      cpu_core/topdown-be-bound/       #     45,5% Backend Bound              (38,85%)
     3.612.581.335      cpu_core/topdown-heavy-ops/      #      2,7% Heavy Operations          #     40,3% Light Operations           (38,85%)
     8.319.156.615      cpu_core/topdown-br-mispredict/  #      6,3% Branch Mispredict         #      0,9% Machine Clears             (38,85%)
     2.631.928.038      cpu_core/topdown-fetch-lat/      #      2,0% Fetch Latency             #      2,3% Fetch Bandwidth            (38,85%)
    27.896.206.958      cpu_core/topdown-mem-bound/      #     21,2% Memory Bound              #     24,3% Core Bound                 (38,85%)

       0,625867225 seconds time elapsed

       4,535275000 seconds user
       1,198751000 seconds sys
 ```